[ Layer ] Move the Weight Read Function to Layer object
It is more flexible to moe Weight Read Fuction to Layer Object itself,
rather than having layer_node. Sometimes, it is necessary to change
the read weight depending on Layer Type, such as Batch normalization
layer.

This PR move general read implementation to Layer_devel and layer is
able to override it as it needs.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: jijoong.moon <jijoong.moon@samsung.com>