[Tensor] Proposal for unifying tensor addition: remove add_i implemenation
This pull request proposes the removal of the add_i implementation and its unification with the add function. Currently, the use of two distinct kernels for tensor addition can result in performance inefficiencies. To enhance overall performance, it is crucial to address this issue. Additionally, we will need to undertake follow-up work to successfully integrate ele_add and sxapy.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>