[CharTensor] QINT8 addition feature
This pull request addresses the activation of the QINT8 element-wise addition feature in CharTensor.
It allows the addition of two tensors of the same dimensions, returning a matrix that contains the sums of the corresponding elements.
Please note that the functionality for automatically determining the new scale factor will be implemented in a future update.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>