Multi-Platform Build Support for Ruy Project in NNTrainer
This pull request enables the Ruy project to be built in NNTrainer across multiple platforms, specifically Ubuntu, Tizen, and Android, and supports both x86 and ARM processors.
For the Tizen and Ubuntu pdebuilds, we are currently using ruy.tar.gz, which contains the packaged Ruy sources, until the Ruy package becomes available.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>