[Tensor] Zero point support for unsigned int type 
This pull request updates the unsigned int type Tensor class to include zero points in its memory management.
It's important to note that the size of the scale factors matches the size of the zero points.
At present, the zero point uses a scale factor size, which will be modified soon.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>