[windows] replace strerror to thread_safe func
In #2870, the thread-safe function strerror_r was replaced
with the non-thread-safe function strerror.

However, due to issues related to thread_safe, it was
re-modified to use the thread-safe function.

For Windows support, a macro was defined to replace
strerror_r with strerror_s only when the operating system is Windows.

**Self evaluation:**
Build test: [x]Passed [ ]Failed [ ]Skipped
Run test: [x]Passed [ ]Failed [ ]Skipped

Signed-off-by: Seungbaek Hong <sb92.hong@samsung.com>