[hgemm] Removal of assert in get_prev_mltpl_of_2p_n
This patch removes the assert statement in the function get_prev_mltpl_of_2p_n, which previously required that the input value N be a multiple of two.
After review, I have identified several instances where N is not necessarily a multiple of two.
Therefore, I've decided to remove this assert statement.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>