[gtest] Remove duplicated main in integration unit test
the test/unittest/integration_tests/integration_tests.cpp
already contains main, causing linking to fail on some Ubuntu 22.04
The suspision is that are those with multilib installed, due to libc difference
The root cause of this issue exposing on some systems is not confirmed.
But both gtest and this compilation unit contain symbol 'main' thus change of
dependency is correct fix.

Signed-off-by: Piotr Rak <piotr.rak@samsung.com>