[ Memory Pool ] Support various buffer size
This PR enables to use mamimum size of buffer for shared tensors.
In nntrainer, we are using shared buffer to use less memory.
For that, memory planner compute the offset and some buffer uses same
offset with diffenent size. It is because we are allocate whole memory
buffer and use it for all tensors. But current rpcmem use allocation
of each tensor and buffer need to allocate new memory block for
different size.
This pr enables compute the maximum size of shared buffer for same
offset and use it thoroughout the process.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: jijoong.moon <jijoong.moon@samsung.com>