[Refactoring] Rename swap -> fsu To prevent confusion
Currently, the terms "FSU" and "memory_swap" are being used together during development, which may cause confusion for both developers and users.
Additionally, considering maintainability of development, it would be better to have a unified name.
- swap keyword also used in std::swap -> it can be make confused too
Previously, it was the Memory Swap function in Training, but now that we are also dealing with inference using Flash Memory, we would like to change the existing term "Swap" to "FSU".

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghak PARK <donghak.park@samsung.com>