OpenCL Context Registration in Engine Class for NNTrainer GPU Support
This pull request addresses a bug in the Engine class related to the registration of the OpenCL context. The newly introduced Engine class lacked the functionality to register the OpenCL context, which prevented NNTrainer from utilizing the GPU engine. Additionally, I corrected a typo that caused the registerContext function to fail in checking the GPU context, preventing it from registering properly with the engine. I have also updated the functions for context registration.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>