Bug Fix: GPU Layer Object Creation in OpenCL Context
This commit addresses an issue related to the creation of a GPU Layer object. With the recent addition of the Engine and Context classes, the Context::createLayerObject function needed to be overridden to successfully create a layer object. Previously, this function returned a nullptr, which ultimately led to a segmentation fault.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>