[ClContext] Streamlining OpenCL Context Management
Previously, there were multiple ClContext instances used for each module—one for the BLAS kernel, one for the layer, and one for the attention kernel. This fragmentation made it difficult to maintain context and registration, potentially leading to undefined behavior and incorrect implementations. To address this issue, I unified the ClContexts into a single instance, which is now part of the Engine object.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>