Bug Fix: OpenCL Kernel Registration Error through Lazy Initialization
This commit fixes an error that occurred during the registration of OpenCL kernels. The issue was caused by the static initialization order, with  in the Engine class being invoked before the kernel strings had been initialized. This patch introduces functions to lazily initialize the kernel strings, ensuring that they are fully initialized before they are used.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>