Add getProperty function
Implemented a getProperty function that allows checking the value using
the key of a set property.
The actual implementation must be added in the layer.
If not implemented, it returns an empty string ("").

To use getProperty, inherit from layer_impl and insert the property tuple
existing in the layer into LayerImpl::for_each function.

If the key exists in the tuple, it returns the value.
If the property is not set and there is no default value, it returns
"empty".

Signed-off-by: SeoHyungjun <hyungjun.seo@samsung.com>