[Feat] Add getProperty function
The getProperty function has been implemented to allow checking values
using the property key.

- For layers inheriting from layer_impl, it can be used directly.
- For layers inheriting from layer_devel, getProperty must be overridden
, and the implementation should enable iterating through the properties
in the layer.

Since properties are stored as tuples, a utility function has been added
in util_func.h to traverse the tuple, compare key values, and return the
corresponding value if found.

Behavior of getProperty:
- If the key exists in the property, it returns the associated value.
- If the property exists but is not set, it returns "empty".
- If the key does not exist in the property, it returns "" (an empty string).

Signed-off-by: SeoHyungjun <hyungjun.seo@samsung.com>