[Tensor][Read] Change Tensor read from file_offset
Previously, the tensor would receive parameters and, if the flag to read from the file offset was true, it would read from the passed start_offset.
Now, since the tensor holds its own fileOffset, if the value is size_t::max, it will default to reading from the stored fileOffset.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghak PARK <donghak.park@samsung.com>