[OpenCL] Enhance Q4_K GEMM Kernel Performance By Vectorization
This commit updates the Q4_K GEMM kernel to take advantage of vectorization.
The inner loop has been modified to process eight contiguous bytes, and a manual reduction is used to collapse the vector into a scalar value.
As a result, this patch enhances the overall performance of the kernel by 2 times.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>