[onnx] add llama modeling script & onnx exporter
When directly exporting Huggingface's LLaMA model to ONNX, it includes conditional statements, loops, and other elements that are not suitable for on-device execution, making it difficult to load in NNTrainer. Therefore, I added a simplest version of the llama modeling file that is compatible with HuggingFace's config and weights while being loadable in NNTrainer. Using this file, you can export HuggingFace's LLaMA model to an ONNX file compatible with NNTrainer.

Some of the changes made are as follows:
- Minimized dependencies on various HuggingFace utilities(inheritance structure..)
  (Caching and sampling are not suitable to be handled within the ONNX model graph)
- Replace all dynamic dimensions with static dimensions
- Instead of including RotaryEmbedding in the model structure, pre-calculated cos/sin values are passed as inputs to the model
- Separated the rsqrt operation in RMSNorm into separate sqrt and pow operations (onnx doesn't support rsqrt)

The following features need to be added to support token generation:
- Group Query Attention
- K/V Cache as inputs
- Causal mask as inputs
(Adding these features is very simple, but they cannot be added yet due to some issues where the graph cannot be loaded in NNTrainer for now..)

Using these files, the graph of the "42dot/42dot_LLM-SFT-1.3B" model on Hugging Face was exported to an ONNX file and loaded in NNTrainer with the following results. (The corresponding changes to NNTrainer will upload in a separate PR.)

```
================================================================================
          Layer name          Layer type    Output dimension         Input layer
================================================================================
         onnx__add_6               input             1:1:1:1
--------------------------------------------------------------------------------
 onnx__add_6/generat            multiout             1:1:1:1         onnx__add_6
--------------------------------------------------------------------------------
                 sin               input            1:1:1:64
--------------------------------------------------------------------------------
 sin/generated_out_0            multiout            1:1:1:64                 sin
--------------------------------------------------------------------------------

...

--------------------------------------------------------------------------------
   model_norm_cast_1                cast          1:1:1:2048      model_norm_mul
--------------------------------------------------------------------------------
    model_norm_mul_1            multiply          1:1:1:2048   model_norm_weight
                                                              model_norm_cast_1
--------------------------------------------------------------------------------
      lm_head_matmul              matmul         1:1:1:50304    model_norm_mul_1
                                                               onnx__matmul_3531
================================================================================
```
(Approximately 8,800 lines, omitted; the total number of layers is estimated to be around 4,000.)

Signed-off-by: Seungbaek Hong <sb92.hong@samsung.com>