[OpenCL] Unpack q4_0x8 into two arrays, scales and quants
Since GPU GEMM kernel takes scales and quants as
input, this adds two functions to unpack q4_0x8
blocks.

- nntrainer::convert_st (single thread version), and
- nntrainer::convert_omp (multi thread version)

`convert_omp` had the better results with unittests. For example,
for a matrix with K, M = 3072, 3072 @ Intel Core Ultra 9 185H 2.50Ghz

- convert_st : 130.812 us
- convert_omp : 17.125 us

In addition,

- gemm_q4_0_cl2

is added. It works the same as gemm_q4_0_cl, but its arguments are
scales and quants. Unpacking is supposed to be done before invoking this
function. Refer to `test/unittest/unittest_cl_helper.cpp` for usage
example.

Signed-off-by: Daekyoung Jung <dk11.jung@samsung.com>