Restricting OpenCL Context Usage When Unavailable
This PR addresses issues where the OpenCL context only considered scenarios in which OpenCL was available.
As a result, there was no proper handling of cases where OpenCL was not available.
This patch now restricts the use of the OpenCL context accordingly.
Please note that additional work may still be required.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>