[tensor] update batchedDot
`Dot` operation was implemented with a focus on width and height,
so it only worked when the channel size was 1.

To support case where the chennal size is greater than 1,
I modified batchedDot function to perform the dot operation repeatedly
not only along the batch dimension but also along the channel dimension.

(the number of `dot` function calls becomes batch * channel,
so it isn't ideal solution, but the current `dot` function cannot handle
the channel, height, and width simultaneously)

**Self evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test: [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: SeungBaek Hong <sb92.hong@samsung.com>