Disable log on Windows
This pull request disables the logging functionality for Windows.
There's an issue with the Windows application in creating and accessing the log file directory using nntrainer_logger.
As a result, I have decided to turn off the logging feature on Windows.

If anyone has a better solution for logging that does not involve nntrainer_logger, please let me know.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Donghyeon Jeong <dhyeon.jeong@samsung.com>